Paint Canvas added to CanvasService
CanvasService is now able to colour the canvas.
Service must be sent appropriate 'move' data, as well as the user id and their permission level. If appropriate, the user will have their wallet deducted before proceeding.
Painting the canvas means creating a record in DrawingHistory and updating a record in Drawing tables.